Moved black and pre-commit to be optional
Currently when building (or including tt-tools-common as a dependency)
it always install blacks and pre-commit, this means that they will
conflict with any packages which also include different versions in
their own dependencies. black and pre-commit are not required for
including the repo as a dependency so they can be moved to be optional
dependencies.

This does change how the repo has to be installed to develop on the
package.

Now to install the dev version you must run `pip install .[dev]`

I also updated the README.md to reflect this change.