Added the option to reset the m3 alongside the chip
In order to keep an m3 update from keeping the board from properly
initializing on the firsst flash we can set arg0 to 3 which will reset
the m3 making sure the new firmware is loaded before the next reset.
This change adds a backwards compatible paramter which allows me to
trigger the m3+asic reset from tt-flash.