Switch required pydantic versioning to accomodate v2
Pydantic has a ground up re-write that's starting to make it into newer
distros, F40, OpenSuse Tumbleweed, etc which includes pydantic v1
support, but breaks the default importing we are doing.

This adjusts the base pydantic in pyproject.toml to be a minimum of 1.2
(to meet Ubuntu 20.04 baseline requirements) and adds a try/except block
to try loading the correct pydantic version where we are expecting.

Quick check on F40, in a venv, this plus the change in tt-tools-common
to do roughly the same thing should sort out tenstorrent/tt-smi#27

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>