Fixing older setuptools again
Ubuntu 22's older setuptools needs setup.py: fine.  However it doesn't
read in things like the fact that we have a binary hanging out and need
that pulled in, like it does in pyprojects.toml.  This sorts that.

It also adds in python3-toml to deal with setup.py pulling in the
version number

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>