ChipReset: Run SBR first, then chip reset
In tt-kmd 2.4.1, secondary bus reset was removed from the ASIC reset code path,
so now it must be performed before reset for each device on the tools side.

This is to make the time between each chip's reset as small as possible to
reduce ethernet timing issues.

Also, create the bdf list before issuing ioctl resets.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>