Remove height styling from TTHostCompatibilityMenu
This widget's render function tracked the number of lines in the widget
based on the number of values added to the text, but sometimes very long
values cause the text to wrap, so the num_lines count was undercounting
the true number of lines and the height of the widget was too short to
display all the data.

If we get rid of this, the widget expands to fill its container, but
this doesn't look too bad.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>